Refactor DependencyService into game module with BaseService architecture
